--- dms-core/src/main/kotlin/team/aliens/dms/domain/remain/dto/QueryCurrentAppliedRemainOptionResponse.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.remain.dto
-
-data class QueryCurrentAppliedRemainOptionResponse(
-    val title: String
-)
