+++ dms-core/src/main/kotlin/team/aliens/dms/domain/remain/dto/RemainResponse.kt
@@ -1,0 +1,40 @@
+package team.aliens.dms.domain.remain.dto
+
+import team.aliens.dms.domain.remain.model.RemainOption
+import java.time.DayOfWeek
+import java.time.LocalTime
+import java.util.UUID
+
+data class RemainOptionsResponse(
+    val remainOptions: List<RemainOptionElement>
+) {
+    data class RemainOptionElement(
+        val id: UUID,
+        val title: String,
+        val description: String,
+        val isApplied: Boolean
+    )
+    companion object {
+        fun of(remainOptions: List<RemainOption>, remainOptionId: UUID?) = RemainOptionsResponse(
+            remainOptions = remainOptions.map {
+                RemainOptionElement(
+                    id = it.id,
+                    title = it.title,
+                    description = it.description,
+                    isApplied = it.id == remainOptionId
+                )
+            }
+        )
+    }
+}
+
+data class RemainAvailableTimeResponse(
+    val startDayOfWeek: DayOfWeek,
+    val startTime: LocalTime,
+    val endDayOfWeek: DayOfWeek,
+    val endTime: LocalTime
+)
+
+data class CurrentAppliedRemainOptionResponse(
+    val title: String
+)
